Move the methods to build the json on another class
!!! Need to add tests on the builder class to verify the built json

(cherry picked from commit a2203547326b940e75aa99b5e2cd43091f0a820a)
